Refactor: pointDetailResolver 통합
