--- src/mocks/point-detail/charges/index.ts
@@ -1,6 +1,0 @@
-import { HttpResponse } from 'msw';
-import pointDetailChargesData from '@assets/data/pointDetailChargesData.json';
-
-export const getPointDetailChargesResolver = () => {
-  return HttpResponse.json(pointDetailChargesData, { status: 200 });
-};
