--- src/mocks/point-detail/total/index.ts
@@ -1,6 +1,0 @@
-import { HttpResponse } from 'msw';
-import PointDetailData from '@assets/data/pointDetailData.json';
-
-export const getPointDetailTotalResolver = () => {
-  return HttpResponse.json(PointDetailData, { status: 200 });
-};
