--- src/mocks/point-detail/usage/index.ts
@@ -1,6 +1,0 @@
-import { HttpResponse } from 'msw';
-import pointDetailUsageData from '@assets/data/pointDetailUsageData.json';
-
-export const getPointDetailUsageResolver = () => {
-  return HttpResponse.json(pointDetailUsageData, { status: 200 });
-};
